+++ src/models/user-preferences/index.ts
@@ -1,0 +1,12 @@
+/*************************************************************************
+ * Copyright 2021 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export { RawUserPreferencesWithMetadata } from './raw-user-preferences-with-metadata';
+export { toUserPreferencesWithMetadata } from './to-user-preferences-with-metadata';
+export { UserPreferences } from './user-preferences';
+export { UserPreferencesWithMetadata } from './user-preferences-with-metadata';
